pyproject.toml
[project]
name = "nishi-toy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.4"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "supabase>=2.28.0",
    "tqdm>=4.67.3",
]
